backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
deepface==0.0.98
distro==1.9.0
filelock==3.24.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
gdown==5.2.1
greenlet==3.3.2
groq==1.0.0
gunicorn==25.1.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
numpy==2.4.2
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
protobuf==7.34.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
retina-face==0.0.17
rich==14.3.3
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.47
sympy==1.14.0
tf_keras==2.20.1
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.1
yarl==1.22.0